AIProcessingService/requirements.txt
grpcio
grpcio-tools
# openai-whisper
google-cloud-speech
librosa
numpy
# transformers
# torchaudio
load_dotenv
